requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
fastapi==0.115.14
h11==0.16.0
idna==3.10
numpy==2.3.1
oracledb==3.2.0
pandas==2.3.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.35.0
minio==7.2.7
